Isotates jumps to their own thread
Isolates jumps forward and backward through the program to run in an
isolated asynchronous thread. This allows the user to abort the jump
in case the computation is too long or even nonhalting. While the
computation is running, the Debugger placed in a ProcessingMode that
only accepts commands to abort the computation.