Changes the Monad for IO commands
The original monad was just IO so impure commands were
    Debugger -> IO Debugger
This was changed to the ErrorIO monad so they are now
    Debugger -> ErrorIO Debugger
This should hopefully make error handling of commands much cleaner.
The change requires that a ErrorIO handler be added to handleCommand
in the Router module to translate IO (Either ErrString Debugger)
values to IO Debugger values. All tests passed, and the <load> and
<write> commands appear to work as expected.