Centralizes file path-finding in tests
Input, output and BF script files used for testing are kept in the
tests/files directory, and they should not be changed. When running
tests, however, you may need to modify a file, so we create a
tests/temp directory that gets removed after the test are completed.
Previously, the paths to these two directories were hard coded in.
Now, the paths are supplied using dedicatde functions so you don't
need to worry about the paths and just the file names. Furthermore,
creating a new mocked debugger for testing automatically finds the
correct test files in their own directory and resets the paths to
the temporary directory so you don't have to do this manually and
potentially overwrite the test files.