Adds the unsafeEdit setting
This allows the user to edit BF scripts when the point of execution
and the point of edit are inside the same while-loop so long as the
point of execution has not advanced beyond the point of edit in the
BF script. Two setting toggles <safe> and <unsafe> are provided for
changing this setting.